+++ src/define.hpp
@@ -98,6 +98,17 @@ bool define_parse(LineData* line)
                     it->first = TokenType::Delete;
                     it++;
                 }
+                // nametable に定義名を追加
+                addNameTable(name, line);
+                // ドット付きの場合は左辺が未登録なら登録
+                auto dot = name.find('.');
+                if (-1 != dot) {
+                    auto left = name.substr(0, dot);
+                    if (nameTable.find(left) == nameTable.end()) {
+                        addNameTable(left, line);
+                        addNameTable(left + ".", line);
+                    }
+                }
             }
             return true;
         }
