+++ src/mnemonic.h
@@ -24,88 +24,6 @@ class TempAddr
 
 extern std::vector<TempAddr*> tempAddrs;
 
-enum class Mnemonic {
-    None = 0,
-    LD,
-    LDI,
-    LDD,
-    LDIR,
-    LDDR,
-    PUSH,
-    POP,
-    EX,
-    EXX,
-    CP,
-    CPI,
-    CPIR,
-    CPD,
-    CPDR,
-    ADD,
-    ADC,
-    SUB,
-    SBC,
-    AND,
-    OR,
-    XOR,
-    INC,
-    DEC,
-    DAA,
-    CPL,
-    NEG,
-    CCF,
-    SCF,
-    NOP,
-    HALT,
-    RL,
-    RLA,
-    RLC,
-    RLCA,
-    RR,
-    RRA,
-    RRC,
-    RRCA,
-    SLA,
-    SLL,
-    SRA,
-    SRL,
-    RLD,
-    RRD,
-    BIT,
-    SET,
-    RES,
-    JP,
-    JR,
-    DJNZ,
-    CALL,
-    RET,
-    RETI,
-    RETN,
-    RST,
-    OUT,
-    OUTI,
-    OTIR,
-    OUTD,
-    OTDR,
-    IN,
-    INI,
-    INIR,
-    IND,
-    INDR,
-    DI,
-    EI,
-    IM,
-    DB,
-    DW,
-    MUL,  // VGS-Zero
-    MULS, // VGS-Zero
-    DIV,  // VGS-Zero
-    DIVS, // VGS-Zero
-    MOD,  // VGS-Zero
-    ATN2, // VGS-Zero
-    SIN,  // VGS-Zero
-    COS,  // VGS-Zero
-};
-
 bool mnemonic_format_check(LineData* line, int size, ...);
 bool mnemonic_format_test(LineData* line, int size, ...);
 bool mnemonic_format_begin(LineData* line, int size, ...);
