--- src/mnemonic_table.hpp
@@ -1,128 +1,0 @@
-/**
- * Z80 Assembler for VGS-Zero
- * Copyright (c) 2024, Yoji Suzuki.
- * License under GPLv3: https://github.com/suzukiplan/vgsasm/blob/master/LICENSE.txt
- */
-#pragma once
-#include "common.h"
-#include "mnemonic.h"
-
-std::map<std::string, Mnemonic> mnemonicTable = {
-    {"LD", Mnemonic::LD},
-    {"LDI", Mnemonic::LDI},
-    {"LDIR", Mnemonic::LDIR},
-    {"LDD", Mnemonic::LDD},
-    {"LDDR", Mnemonic::LDDR},
-    {"PUSH", Mnemonic::PUSH},
-    {"POP", Mnemonic::POP},
-    {"EX", Mnemonic::EX},
-    {"EXX", Mnemonic::EXX},
-    {"CP", Mnemonic::CP},
-    {"CPI", Mnemonic::CPI},
-    {"CPIR", Mnemonic::CPIR},
-    {"CPD", Mnemonic::CPD},
-    {"CPDR", Mnemonic::CPDR},
-    {"ADD", Mnemonic::ADD},
-    {"ADC", Mnemonic::ADC},
-    {"SUB", Mnemonic::SUB},
-    {"SBC", Mnemonic::SBC},
-    {"AND", Mnemonic::AND},
-    {"OR", Mnemonic::OR},
-    {"XOR", Mnemonic::XOR},
-    {"EOR", Mnemonic::XOR},
-    {"INC", Mnemonic::INC},
-    {"DEC", Mnemonic::DEC},
-    {"DAA", Mnemonic::DAA},
-    {"CPL", Mnemonic::CPL},
-    {"NEG", Mnemonic::NEG},
-    {"CCF", Mnemonic::CCF},
-    {"SCF", Mnemonic::SCF},
-    {"NOP", Mnemonic::NOP},
-    {"HALT", Mnemonic::HALT},
-    {"RL", Mnemonic::RL},
-    {"RLA", Mnemonic::RLA},
-    {"RLC", Mnemonic::RLC},
-    {"RLCA", Mnemonic::RLCA},
-    {"RR", Mnemonic::RR},
-    {"RRA", Mnemonic::RRA},
-    {"RRC", Mnemonic::RRC},
-    {"RRCA", Mnemonic::RRCA},
-    {"SLA", Mnemonic::SLA},
-    {"SL", Mnemonic::SLA},
-    {"SLL", Mnemonic::SLL},
-    {"SRA", Mnemonic::SRA},
-    {"SRL", Mnemonic::SRL},
-    {"SR", Mnemonic::SRL},
-    {"RLD", Mnemonic::RLD},
-    {"RRD", Mnemonic::RRD},
-    {"BIT", Mnemonic::BIT},
-    {"SET", Mnemonic::SET},
-    {"RES", Mnemonic::RES},
-    {"JP", Mnemonic::JP},
-    {"JR", Mnemonic::JR},
-    {"DJNZ", Mnemonic::DJNZ},
-    {"CALL", Mnemonic::CALL},
-    {"RST", Mnemonic::RST},
-    {"RET", Mnemonic::RET},
-    {"RETI", Mnemonic::RETI},
-    {"RETN", Mnemonic::RETN},
-    {"OUT", Mnemonic::OUT},
-    {"OUTI", Mnemonic::OUTI},
-    {"OUTIR", Mnemonic::OTIR},
-    {"OTIR", Mnemonic::OTIR},
-    {"OUTD", Mnemonic::OUTD},
-    {"OUTDR", Mnemonic::OTDR},
-    {"OTDR", Mnemonic::OTDR},
-    {"IN", Mnemonic::IN},
-    {"INI", Mnemonic::INI},
-    {"INIR", Mnemonic::INIR},
-    {"IND", Mnemonic::IND},
-    {"INDR", Mnemonic::INDR},
-    {"DI", Mnemonic::DI},
-    {"EI", Mnemonic::EI},
-    {"IM", Mnemonic::IM},
-    {"DB", Mnemonic::DB},
-    {"DEFB", Mnemonic::DB},
-    {"DW", Mnemonic::DW},
-    {"DEFW", Mnemonic::DW},
-    {"BYTE", Mnemonic::DB},
-    {"WORD", Mnemonic::DW},
-    {"MUL", Mnemonic::MUL},
-    {"MULS", Mnemonic::MULS},
-    {"DIV", Mnemonic::DIV},
-    {"DIVS", Mnemonic::DIVS},
-    {"MOD", Mnemonic::MOD},
-    {"ATN2", Mnemonic::ATN2},
-    {"SIN", Mnemonic::SIN},
-    {"COS", Mnemonic::COS},
-};
-
-std::map<std::string, Operand> operandTable = {
-    {"A", Operand::A},
-    {"B", Operand::B},
-    {"C", Operand::C},
-    {"D", Operand::D},
-    {"E", Operand::E},
-    {"F", Operand::F},
-    {"H", Operand::H},
-    {"L", Operand::L},
-    {"IXH", Operand::IXH},
-    {"IXL", Operand::IXL},
-    {"IYH", Operand::IYH},
-    {"IYL", Operand::IYL},
-    {"AF", Operand::AF},
-    {"AF'", Operand::AF_dash},
-    {"BC", Operand::BC},
-    {"DE", Operand::DE},
-    {"HL", Operand::HL},
-    {"IX", Operand::IX},
-    {"IY", Operand::IY},
-    {"SP", Operand::SP},
-    {"NZ", Operand::NZ},
-    {"Z", Operand::Z},
-    {"NC", Operand::NC},
-    {"PO", Operand::PO},
-    {"PE", Operand::PE},
-    {"P", Operand::P},
-    {"M", Operand::M},
-};
